requirements.txt
jinja2>=3.1.0
PyYAML>=6.0.0
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9
